requirements.txt
arxiv
requests
scholarly
huggingface_hub
PyGithub
openai
anthropic
google-generativeai
pyyaml
matplotlib
pandas
numpy
scikit-learn
torch
pylatex
subprocess32
python-dotenv
